FIX: Deletion of existing annotation text is suppressed if the id is non-numeric because it is new [t26756]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@16668 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ plugins/annotate/pages/save.php
@@ -33,7 +33,10 @@ if ($oldtext!="")
 	remove_keyword_mappings($ref,i18n_get_indexable($oldtext),-1,false,false,"annotation_ref",$id);
     }
 
-sql_query("delete from annotate_notes where ref='$ref' and note_id='$id'");
+if (ctype_digit($id))
+    {
+    sql_query("delete from annotate_notes where ref='$ref' and note_id='$id'");
+    }
 
 if (substr($text,0,strlen($username))!=$username)
     {
